feat: Add suspense
